Updated exports to match docs
